concmon: extract from kvserver/concurrency
Nathan wrote this nifty test helper that can help orchestrating tests
for subsystems that have multiple goroutines, and getting deterministic
datadriven trace-driven output.

I have something I'm hoping to use this for soon (roachtest randomized
testing), so pull it out into its own package.

Release note: None
